fix: rename modules
