+++ hstream/hstream.cabal
@@ -26,7 +26,7 @@ source-repository head
 
 library
   exposed-modules:
-    HStream.Server.Common
+    HStream.Server.Handler.Common
     HStream.Server.Exception
     HStream.Server.Handler
     HStream.Server.Persistence
