+++ hstream/src/HStream/Server/Handler.hs
@@ -58,9 +58,9 @@ import           HStream.Processing.Processor      (TaskBuilder, getTaskName,
                                                     runTask)
 import           HStream.Processing.Type
 import           HStream.SQL.Codegen
-import           HStream.Server.Common
 import           HStream.Server.Exception
 import           HStream.Server.HStreamApi
+import           HStream.Server.Handler.Common
 import qualified HStream.Server.Persistence        as P
 import           HStream.Store                     hiding (e)
 import           HStream.Utils
